# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/posts": {
      "post": {
        "operationId": "PostController_createPost_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo bài đăng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không có quyền truy cập"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo bài đăng mới",
        "tags": [
          "Post"
        ]
      },
      "get": {
        "operationId": "PostController_getNewsfeed_v1",
        "parameters": [
          {
            "name": "sport",
            "required": false,
            "in": "query",
            "description": "Lọc theo môn thể thao",
            "schema": {
              "enum": [
                "football",
                "basketball",
                "tennis",
                "badminton",
                "volleyball",
                "table_tennis",
                "swimming",
                "running",
                "cycling",
                "yoga",
                "gym",
                "martial_arts",
                "boxing",
                "golf",
                "rock_climbing"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bài đăng trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bài đăng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách bài đăng (newsfeed) của user hiện tại",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/all": {
      "get": {
        "operationId": "PostController_getAllPosts_v1",
        "parameters": [
          {
            "name": "sport",
            "required": false,
            "in": "query",
            "description": "Lọc theo môn thể thao",
            "schema": {
              "enum": [
                "football",
                "basketball",
                "tennis",
                "badminton",
                "volleyball",
                "table_tennis",
                "swimming",
                "running",
                "cycling",
                "yoga",
                "gym",
                "martial_arts",
                "boxing",
                "golf",
                "rock_climbing"
              ],
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Lọc theo ID người dùng",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bài đăng trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy tất cả bài đăng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Lấy tất cả bài đăng",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/{postId}": {
      "get": {
        "operationId": "PostController_getPostById_v1",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "ID của bài đăng",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin bài đăng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bài đăng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy thông tin chi tiết bài đăng theo ID",
        "tags": [
          "Post"
        ]
      },
      "put": {
        "operationId": "PostController_updatePost_v1",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "ID của bài đăng",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật bài đăng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không có quyền truy cập"
          },
          "403": {
            "description": "Không có quyền chỉnh sửa bài đăng này"
          },
          "404": {
            "description": "Không tìm thấy bài đăng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật bài đăng",
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_deletePost_v1",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "ID của bài đăng",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bài đăng thành công"
          },
          "401": {
            "description": "Không có quyền truy cập"
          },
          "403": {
            "description": "Không có quyền xóa bài đăng này"
          },
          "404": {
            "description": "Không tìm thấy bài đăng"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa bài đăng",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/user/{userId}": {
      "get": {
        "operationId": "PostController_getPostsByUserId_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID của người dùng",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bài đăng trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bài đăng của người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách bài đăng của người dùng theo ID",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/hashtags/trending": {
      "get": {
        "operationId": "PostController_getTrendingHashtags_v1",
        "parameters": [
          {
            "name": "timeRange",
            "required": false,
            "in": "query",
            "description": "Khoảng thời gian (7d, 30d, 90d, all)",
            "schema": {
              "example": "7d",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng hashtag trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách hashtag trending thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendingHashtagsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Lấy danh sách hashtag trending",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/v1/posts/hashtags/{hashtag}/posts": {
      "get": {
        "operationId": "PostController_getPostsByHashtag_v1",
        "parameters": [
          {
            "name": "hashtag",
            "required": true,
            "in": "path",
            "description": "Tên hashtag (có thể có hoặc không có dấu #)",
            "schema": {
              "example": "fitness",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số lượng bài đăng trên mỗi trang (mặc định: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Số trang (mặc định: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bài đăng theo hashtag thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy bài đăng với hashtag này"
          }
        },
        "summary": "Lấy danh sách bài đăng theo hashtag",
        "tags": [
          "Post"
        ]
      }
    },
...
    "/api/v1/posts/{postId}/clear-url": {
      "delete": {
        "operationId": "PostController_clearUrl_v1",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "ID của bài đăng",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Xóa thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không có quyền truy cập"
          },
          "403": {
            "description": "Không có quyền chỉnh sửa bài đăng này"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa images và video của bài đăng (clearUrl)",
        "tags": [
          "Post"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger docs for the project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginatedPostsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Trang hiện tại",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Số lượng items trên mỗi trang",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "description": "Tổng số trang",
            "example": 10
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Có trang tiếp theo không",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "description": "Có trang trước không",
            "example": false
          },
          "posts": {
            "description": "Danh sách bài đăng",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostResponseDto"
            }
          }
        },
        "required": [
          "page",
          "limit",
          "totalPages",
          "hasNextPage",
          "hasPrevPage",
          "posts"
        ]
      },
...
      "ClearUrlDto": {
        "type": "object",
        "properties": {
          "isClearImage": {
            "type": "boolean",
            "description": "Xóa images",
            "default": false
          },
          "isClearVideo": {
            "type": "boolean",
            "description": "Xóa video",
            "default": false
          }
        }
      },
...
    }
  }
}